requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
Pillow==9.3.0
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.6
